+++ src/tikray/macropipe/__init__.py
@@ -1,0 +1,8 @@
+import tikray.macropipe.lib  # noqa: F401
+
+from .core import MacroPipe, recipe
+
+__all__ = [
+    "recipe",
+    "MacroPipe",
+]